@@ -2,20 +2,17 @@
   "name": "deluge-rpc-socket",
   "version": "0.4.0",
   "description": "Node.js API for Deluge's Binary Socket RPC API",
-  "main": "src/DelugeRPC",
-  "types": "src/DelugeRPC.d.ts",
+  "main": "dist/DelugeRPC.js",
+  "types": "dist/DelugeRPC.d.ts",
   "scripts": {
-    "format": "prettier --write \"**/*.@(ts|js?(on)|md|prettierrc|clang-format|yml)\"",
-    "ci": "yarn test && yarn coveralls",
-    "test": "jest --passWithNoTests",
-    "coveralls": "coveralls < coverage/lcov.info",
-    "preversion": "yarn format",
-    "postversion": "npm publish",
+    "format": "prettier **/*",
+    "spellcheck": "cspell **/*",
+    "test": "jest",
+    "preversion": "yarn format --write",
     "build": "tsc --declaration",
-    "clean": "rm -f src/DelugeRPC.{d.ts,js}",
+    "clean": "rm -f dist",
     "checks": "yarn format && yarn build && yarn test",
-    "prepublishOnly": "yarn build",
-    "postpublish": "git push origin master --tags",
+    "prepare": "yarn build",
     "start": "yarn repl",
     "prerepl": "yarn clean",
     "repl": "ts-node repl.ts --experimental-repl-await",
@@ -23,31 +20,34 @@
     "mkconfig": "echo Making new config.js for testing purposes. Please update it manually. && cp -i config.sample.js config.js"
   },
   "files": [
-    "src/DelugeRPC.ts",
-    "src/DelugeRPC.js",
-    "src/DelugeRPC.d.ts"
+    "dist/DelugeRPC.js",
+    "dist/DelugeRPC.d.ts",
+    "dist/utils/SharedPromise.js"
   ],
   "repository": "git@github.com:cinderblock/node-deluge-rpc.git",
   "author": "Cameron Tacklind <cameron@tacklind.com>",
   "license": "MIT",
   "dependencies": {
     "camelcase-keys-deep": "^0.1.0",
-    "pako": "^1.0.6",
-    "python-rencode": "^1.3.2",
-    "smallest-power-of-two": "^1.0.0",
-    "snakecase-keys": "^2.1.0"
+    "pako": "^1.0.10",
+    "python-rencode": "^1.4.4",
+    "smallest-power-of-two": "^1.1.2",
+    "snakecase-keys": "^3.1.0"
   },
   "devDependencies": {
-    "@types/jest": "^23.3.9",
-    "@types/node": "^10.12.9",
-    "coveralls": "^3.0.2",
+    "@types/camelcase-keys-deep": "^0.1.0",
+    "@types/jest": "^24.0.25",
+    "@types/node": "^13.1.2",
+    "@types/pako": "^1.0.1",
+    "coveralls": "^3.0.9",
+    "cspell": "^4.0.43",
     "install-vscode-extensions": "^1.0.0",
-    "jest": "^23.6.0",
-    "prettier": "^1.15.2",
+    "jest": "^24.9.0",
+    "prettier": "^1.19.1",
     "repl-live": "^0.1.0",
-    "ts-jest": "^23.10.4",
-    "ts-node": "^7.0.1",
-    "typescript": "^3.1.6",
-    "yarn": "^1.12.3"
+    "ts-jest": "^24.2.0",
+    "ts-node": "^8.5.4",
+    "typescript": "^3.7.4",
+    "yarn": "^1.21.1"
   }
 }
